use a periodic timer to drive the car and let the main loop for looking around

+++ SelfDrivingCar/ServoExt.cpp
@@ -1,15 +1,13 @@
 #include "ServoExt.h"
 
-#include "Arduino.h"
-
-ServoExt::ServoExt(const unsigned short int maxAngle, const unsigned short int travelTime) {
+ServoExt::ServoExt(const byte maxAngle, const int travelTime) {
   this->maxAngle = maxAngle;
   this->travelTime = travelTime;
   this->toStabilize = 0;
   this->prevAngle = this->oldAngle = 90;
 }
 
-void ServoExt::setAngle(unsigned short int angle, const bool waitToStabilize) {
+void ServoExt::setAngle(byte angle, const bool waitToStabilize) {
   angle = max(angle, 90 - this->maxAngle);
   angle = min(angle, 90 + this->maxAngle);
 
@@ -47,7 +45,7 @@ void ServoExt::setAngle(unsigned short int angle, const bool waitToStabilize) {
   }
 }
 
-void ServoExt::setPercentage(const unsigned short int percentage, const bool waitToStabilize) {
+void ServoExt::setPercentage(const int percentage, const bool waitToStabilize) {
   setAngle(90 + percentage * maxAngle / 100, waitToStabilize);
 }
 
@@ -56,6 +54,6 @@ void ServoExt::waitForAngle() {
     delay(millis() - toStabilize);
 }
 
-unsigned short int ServoExt::getAngle() {
+byte ServoExt::getAngle() {
   return this->prevAngle;
 }
